docs: Update the metadata format (#48)
* docs: Update the metadata format

Update the documentation and JSON Schema following metadata format
changes. Notably:

1. Wheel metadata uses `variant.json` instead of Core Metadata
   extensions now.

2. Default priorities are specified within namespaces and features
   now rather than globally.

While at it, take the opportunity to streamline the format much like
we streamlined the parsers in `variantlib`. All three files use the same
structure, with the same key names, so describe them once instead of
repeating for every file. Be more specific about the tree structure
and data types. Limit the file descriptions to their unique properties
and examples.

* docs: Update for the plugin-api changes

Update for the following changes:

1. `plugin-api` can now specify non-callable

2. `plugin-api` can reference the whole module (like PEP 517)

3. `plugin-api` is optional and is inferred from `requires` when missing

* docs: Add a `variantlib generate-index-json` example